Refactor location of displays to fit with driver/mock layout
